+++ ft_case_x_big.c
@@ -6,13 +6,13 @@
 /*   By: glaurent <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/14 01:17:23 by glaurent          #+#    #+#             */
-/*   Updated: 2019/11/14 11:13:04 by glaurent         ###   ########.fr       */
+/*   Updated: 2019/11/14 11:28:38 by glaurent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_width_num_x_big(t_struct *p, unsigned int copy, char c)
+void	ft_width_num_x_big(t_struct *p, unsigned long copy, char c)
 {
    int     i;
