+++ ft_utils.c
@@ -6,7 +6,7 @@
 /*   By: gaefourn <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/10 23:13:17 by gaefourn          #+#    #+#             */
-/*   Updated: 2019/11/14 05:55:17 by gaefourn         ###   ########.fr       */
+/*   Updated: 2019/11/14 11:21:31 by glaurent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void	ft_putstr(char *str, t_struct *p)
 
 void	ft_putnbr_base(unsigned long long n, char *base, t_struct *p)
 {
-	int					i;
+	unsigned long long	i;
 
 	i = 0;
 	while (base[i])
